changed type of arguments for controls to int & store controls in slice, fixed...

See merge request hubman/midi_manipulator!19

+++ pkg/midi/in.go
@@ -41,9 +41,9 @@ func (md *MidiDevice) getMidiMessage(msg midi.Message, _ int32) {
 		}
 	case msg.GetControlChange(&channel, &key, &velocity):
 		// CONTROL PUSHED STATUS
-		velocity, valid := md.handleControls(key, velocity)
+		velocity, valid := md.handleControls(int(key), int(velocity))
 		if valid {
-			kctx := KeyContext{key: key, velocity: velocity, usedAt: time.Now(),
+			kctx := KeyContext{key: key, velocity: uint8(velocity), usedAt: time.Now(),
 				status: model.ControlPushed{Device: md.name, KeyCode: int(key), Value: int(velocity)}}
 			md.clickBuffer.SetKeyContext(key, kctx)
 		}
